# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Employee {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  dailyWage: number;
  createdAt: string;
}

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  /** @minimum 0 */
  dailyWage: number;
}

export interface EmployeeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @minimum 0 */
  dailyWage?: number;
}

export type AttendanceRecordStatus = typeof AttendanceRecordStatus[keyof typeof AttendanceRecordStatus];


export const AttendanceRecordStatus = {
  present: 'present',
  absent: 'absent',
  holiday: 'holiday',
  vacation: 'vacation',
} as const;

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  year: number;
  month: number;
  day: number;
  status: AttendanceRecordStatus;
}

export type AttendanceInputStatus = typeof AttendanceInputStatus[keyof typeof AttendanceInputStatus];


export const AttendanceInputStatus = {
  present: 'present',
  absent: 'absent',
  holiday: 'holiday',
  vacation: 'vacation',
} as const;

export interface AttendanceInput {
  employeeId: number;
  year: number;
  month: number;
  day: number;
  status: AttendanceInputStatus;
}

export interface EmployeeSummary {
  employeeId: number;
  employeeName: string;
  /** @nullable */
  phone?: string | null;
  dailyWage: number;
  totalPresent: number;
  totalAbsent: number;
  totalHoliday: number;
  totalVacation: number;
  totalSalary: number;
}

export interface MonthlyStats {
  year: number;
  month: number;
  totalEmployees: number;
  totalPresent: number;
  totalAbsent: number;
  totalHoliday: number;
  totalVacation: number;
  totalSalaryPaid: number;
}

export type ListAttendanceParams = {
employeeId?: number;
year?: number;
month?: number;
};

export type GetAttendanceSummaryParams = {
year: number;
month: number;
};

export type GetMonthlyStatsParams = {
year: number;
month: number;
};