# lib/api-zod/src/generated/types/attendanceInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceInputStatus } from './attendanceInputStatus';

export interface AttendanceInput {
  employeeId: number;
  year: number;
  month: number;
  day: number;
  status: AttendanceInputStatus;
}