# lib/api-zod/src/generated/types/attendanceRecord.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceRecordStatus } from './attendanceRecordStatus';

export interface AttendanceRecord {
  id: number;
  employeeId: number;
  year: number;
  month: number;
  day: number;
  status: AttendanceRecordStatus;
}