# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  name: string;
  /** @nullable */
  phone?: string | null;
  dailyWage: number;
  createdAt: string;
}