# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeInput {
  /** @minLength 1 */
  name: string;
  phone?: string;
  /** @minimum 0 */
  dailyWage: number;
}