# lib/api-zod/src/generated/types/employeeSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeSummary {
  employeeId: number;
  employeeName: string;
  /** @nullable */
  phone?: string | null;
  dailyWage: number;
  totalPresent: number;
  totalAbsent: number;
  totalHoliday: number;
  totalVacation: number;
  totalSalary: number;
}