# lib/api-zod/src/generated/types/employeeUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface EmployeeUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  phone?: string | null;
  /** @minimum 0 */
  dailyWage?: number;
}