# lib/api-zod/src/generated/types/monthlyStats.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Sabah General Trading - Employee Attendance System
 * OpenAPI spec version: 0.1.0
 */

export interface MonthlyStats {
  year: number;
  month: number;
  totalEmployees: number;
  totalPresent: number;
  totalAbsent: number;
  totalHoliday: number;
  totalVacation: number;
  totalSalaryPaid: number;
}